Failing app
In this commit I have moved the responder into it's own module.
I have added the test environment folder to the gitignore
Tried to update the app to work with the proxy pass

Decided to give up on this branch for now and move to a pure console loop based model.